+++ sdk/identity/azure-identity/src/main/java/com/azure/identity/implementation/IdentityClient.java
@@ -18,6 +18,7 @@ import com.azure.identity.implementation.util.LoggingUtil;
 import com.azure.identity.implementation.util.ScopeUtil;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.microsoft.aad.msal4j.AuthorizationCodeParameters;
+import com.microsoft.aad.msal4j.AppTokenProviderParameters;
 import com.microsoft.aad.msal4j.ClaimsRequest;
 import com.microsoft.aad.msal4j.ClientCredentialFactory;
 import com.microsoft.aad.msal4j.ClientCredentialParameters;
@@ -30,13 +31,13 @@ import com.microsoft.aad.msal4j.MsalInteractionRequiredException;
 import com.microsoft.aad.msal4j.PublicClientApplication;
 import com.microsoft.aad.msal4j.RefreshTokenParameters;
 import com.microsoft.aad.msal4j.SilentParameters;
+import com.microsoft.aad.msal4j.TokenProviderResult;
 import com.microsoft.aad.msal4j.UserNamePasswordParameters;
 import com.sun.jna.Platform;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import javax.net.ssl.HttpsURLConnection;
-import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
@@ -62,6 +63,7 @@ import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.function.Supplier;
 
 /**
@@ -72,6 +74,7 @@ public class IdentityClient extends IdentityClientBase {
     private final SynchronizedAccessor<PublicClientApplication> publicClientApplicationAccessor;
     private final SynchronizedAccessor<ConfidentialClientApplication> confidentialClientApplicationAccessor;
     private final SynchronizedAccessor<ConfidentialClientApplication> managedIdentityConfidentialClientApplicationAccessor;
+    private final SynchronizedAccessor<ConfidentialClientApplication> workloadIdentityConfidentialClientApplicationAccessor;
     private final SynchronizedAccessor<String> clientAssertionAccessor;
 
 
@@ -106,6 +109,9 @@ public class IdentityClient extends IdentityClientBase {
         this.managedIdentityConfidentialClientApplicationAccessor =
             new SynchronizedAccessor<>(this::getManagedIdentityConfidentialClientApplication);
 
+        this.workloadIdentityConfidentialClientApplicationAccessor =
+            new SynchronizedAccessor<>(this::getWorkloadIdentityConfidentialClientApplication);
+
         Duration cacheTimeout = (clientAssertionTimeout == null) ? Duration.ofMinutes(5) : clientAssertionTimeout;
         this.clientAssertionAccessor = new SynchronizedAccessor<>(this::parseClientAssertion, cacheTimeout);
     }
@@ -124,6 +130,16 @@ public class IdentityClient extends IdentityClientBase {
         return Mono.defer(() -> {
             try {
                 return Mono.just(super.getManagedIdentityConfidentialClient());
+            } catch (RuntimeException e) {
+                return Mono.error(e);
+            }
+        });
+    }
+
+    private Mono<ConfidentialClientApplication> getWorkloadIdentityConfidentialClientApplication() {
+        return Mono.defer(() -> {
+            try {
+                return Mono.just(super.getWorkloadIdentityConfidentialClient());
             } catch (RuntimeException e) {
                 return Mono.error(e);
             }
@@ -219,7 +235,7 @@ public class IdentityClient extends IdentityClientBase {
                         ConfidentialClientApplication.builder(spDetails.get("client"),
                             ClientCredentialFactory.createFromSecret(spDetails.get("key")))
                             .authority(authorityUrl)
-                            .instanceDiscovery(options.getDisableAuthorityValidationAndInstanceDiscovery());
+                            .instanceDiscovery(options.isInstanceDiscoveryEnabled());
 
                     // If http pipeline is available, then it should override the proxy options if any configured.
                     if (httpPipelineAdapter != null) {
@@ -492,10 +508,6 @@ public class IdentityClient extends IdentityClientBase {
                     builder.clientCredential(ClientCredentialFactory
                         .createFromClientAssertion(clientAssertionSupplier.get()));
                 }
-                if (request.getClaims() != null) {
-                    ClaimsRequest customClaimRequest = CustomClaimRequest.formatAsClaimsRequest(request.getClaims());
-                    builder.claims(customClaimRequest);
-                }
                 return confidentialClient.acquireToken(builder.build());
             }
         )).map(MsalToken::new);
@@ -503,6 +515,19 @@ public class IdentityClient extends IdentityClientBase {
 
     public Mono<AccessToken> authenticateWithManagedIdentityConfidentialClient(TokenRequestContext request) {
         return managedIdentityConfidentialClientApplicationAccessor.getValue()
+            .flatMap(confidentialClient -> Mono.fromFuture(() -> {
+                    ClientCredentialParameters.ClientCredentialParametersBuilder builder =
+                        ClientCredentialParameters.builder(new HashSet<>(request.getScopes()))
+                            .tenant(IdentityUtil
+                                .resolveTenantId(tenantId, request, options));
+                    return confidentialClient.acquireToken(builder.build());
+                }
+            )).onErrorMap(t -> new CredentialUnavailableException("Managed Identity authentication is not available.", t))
+            .map(MsalToken::new);
+    }
+
+    public Mono<AccessToken> authenticateWithWorkloadIdentityConfidentialClient(TokenRequestContext request) {
+        return workloadIdentityConfidentialClientApplicationAccessor.getValue()
             .flatMap(confidentialClient -> Mono.fromFuture(() -> {
                     ClientCredentialParameters.ClientCredentialParametersBuilder builder =
                         ClientCredentialParameters.builder(new HashSet<>(request.getScopes()))
@@ -602,11 +627,6 @@ public class IdentityClient extends IdentityClientBase {
                 SilentParameters.SilentParametersBuilder parametersBuilder = SilentParameters.builder(
                         new HashSet<>(request.getScopes()))
                     .tenant(IdentityUtil.resolveTenantId(tenantId, request, options));
-                if (request.getClaims() != null) {
-                    ClaimsRequest customClaimRequest = CustomClaimRequest
-                        .formatAsClaimsRequest(request.getClaims());
-                    parametersBuilder.claims(customClaimRequest);
-                }
                 try {
                     return confidentialClient.acquireTokenSilently(parametersBuilder.build());
                 } catch (MalformedURLException e) {
@@ -902,41 +922,7 @@ public class IdentityClient extends IdentityClientBase {
     public Mono<AccessToken> authenticateWithExchangeToken(TokenRequestContext request) {
 
         return clientAssertionAccessor.getValue()
-            .flatMap(assertionToken -> Mono.fromCallable(() -> {
-                String authorityUrl = TRAILING_FORWARD_SLASHES.matcher(options.getAuthorityHost()).replaceAll("")
-                    + "/" + tenantId + "/oauth2/v2.0/token";
-
-                String urlParams = "client_assertion=" + assertionToken
-                    + "&client_assertion_type=urn:ietf:params:oauth:client-assertion-type:jwt-bearer&client_id="
-                    + clientId + "&grant_type=client_credentials&scope=" + urlEncode(request.getScopes().get(0));
-
-                byte[] postData = urlParams.getBytes(StandardCharsets.UTF_8);
-                int postDataLength = postData.length;
-
-                HttpURLConnection connection = null;
-
-                URL url = getUrl(authorityUrl);
-
-                try {
-                    connection = (HttpURLConnection) url.openConnection();
-                    connection.setRequestMethod("POST");
-                    connection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
-                    connection.setRequestProperty("Content-Length", Integer.toString(postDataLength));
-                    connection.setRequestProperty("User-Agent", userAgent);
-                    connection.setDoOutput(true);
-                    try (DataOutputStream outputStream = new DataOutputStream(connection.getOutputStream())) {
-                        outputStream.write(postData);
-                    }
-                    connection.connect();
-
-                    return SERIALIZER_ADAPTER.deserialize(connection.getInputStream(), MSIToken.class,
-                        SerializerEncoding.JSON);
-                } finally {
-                    if (connection != null) {
-                        connection.disconnect();
-                    }
-                }
-            }));
+            .flatMap(assertionToken -> Mono.fromCallable(() -> authenticateWithExchangeTokenHelper(request, assertionToken)));
     }
 
     /**
@@ -1088,9 +1074,6 @@ public class IdentityClient extends IdentityClientBase {
         });
     }
 
-    static URL getUrl(String uri) throws MalformedURLException {
-        return new URL(uri);
-    }
     /**
      * Asynchronously acquire a token from the Virtual Machine IMDS endpoint.
      *
@@ -1237,8 +1220,6 @@ public class IdentityClient extends IdentityClientBase {
         }
     }
 
-
-
     void openUrl(String url) throws IOException {
         Runtime rt = Runtime.getRuntime();
 
@@ -1273,7 +1254,22 @@ public class IdentityClient extends IdentityClientBase {
         return ADFS_TENANT.equals(this.tenantId);
     }
 
-    private static String urlEncode(String value) throws IOException {
-        return URLEncoder.encode(value, StandardCharsets.UTF_8.name());
+    Function<AppTokenProviderParameters, CompletableFuture<TokenProviderResult>> getWorkloadIdentityTokenProvider() {
+        return appTokenProviderParameters -> {
+            TokenRequestContext trc = new TokenRequestContext()
+                .setScopes(new ArrayList<>(appTokenProviderParameters.scopes))
+                .setClaims(appTokenProviderParameters.claims)
+                .setTenantId(appTokenProviderParameters.tenantId);
+
+            Mono<AccessToken> accessTokenAsync = authenticateWithExchangeToken(trc);
+
+            return accessTokenAsync.map(accessToken -> {
+                TokenProviderResult result = new TokenProviderResult();
+                result.setAccessToken(accessToken.getToken());
+                result.setTenantId(trc.getTenantId());
+                result.setExpiresInSeconds(accessToken.getExpiresAt().toEpochSecond());
+                return result;
+            }).toFuture();
+        };
     }
 }